workflows: harvested records
* Shows links to harvested records when using OAI harvest
  workflows.

* Tracks any spawned records when using the task "start_async_workflow".
