+++ invenio/modules/oaiharvester/workflows/oaiharvest_harvest_repositories.py
@@ -120,3 +120,77 @@ class oaiharvest_harvest_repositories(RecordWorkflow):
         workflows_reviews(stop_if_error=True),
         update_last_update(get_repositories_list())
     ]
+
+    @staticmethod
+    def get_description(bwo):
+        """Return description of object."""
+        from flask import render_template
+
+        identifiers = None
+
+        extra_data = bwo.get_extra_data()
+        if 'options' in extra_data and 'identifiers' in extra_data["options"]:
+            identifiers = extra_data["options"]["identifiers"]
+
+        results = bwo.get_tasks_results()
+
+        if 'review_workflow' in results:
+            result_progress = results['review_workflow'][0]['result']
+        else:
+            result_progress = {}
+
+        current_task = extra_data['_last_task_name']
+
+        return render_template("workflows/styles/harvesting_description.html",
+                               identifiers=identifiers,
+                               result_progress=result_progress,
+                               current_task=current_task)
+
+    @staticmethod
+    def get_title(bwo):
+        """Return title of object."""
+        return "Summary of OAI harvesting from: {0}".format(
+            bwo.get_extra_data()["repository"]["name"])
+
+    @staticmethod
+    def formatter(bwo, of="hb"):
+        """Return description of object."""
+        from flask import render_template
+        from invenio.modules.workflows.models import BibWorkflowObject
+        from invenio.modules.workflows.registry import workflows
+
+        identifiers = None
+
+        extra_data = bwo.get_extra_data()
+        if 'options' in extra_data and 'identifiers' in extra_data["options"]:
+            identifiers = extra_data["options"]["identifiers"]
+
+        results = bwo.get_tasks_results()
+
+        if 'review_workflow' in results:
+            result_progress = results['review_workflow'][0]['result']
+        else:
+            result_progress = {}
+
+        current_task = extra_data['_last_task_name']
+
+        related_objects = []
+        for id_object in extra_data.get("objects_spawned", list()):
+            spawned_object = BibWorkflowObject.query.get(id_object)
+            if spawned_object:
+                workflow = workflows.get(spawned_object.get_workflow_name())
+                related_objects.append(
+                    (spawned_object.id,
+                     workflow.get_title(spawned_object) or "No title")
+                )
+            else:
+                related_objects.append(
+                    (id_object,
+                     None)
+                )
+
+        return render_template("workflows/styles/harvesting_description.html",
+                               identifiers=identifiers,
+                               result_progress=result_progress,
+                               current_task=current_task,
+                               related_objects=related_objects)
